+++ bloc_architecture_widget/lib/app_params.dart
@@ -8,17 +8,17 @@ class AppParams {
   const AppParams._();
 
   /// Style curve
-  static const Curve animationCurve = Curves.fastOutSlowIn;
+  static Curve animationCurve = Curves.fastOutSlowIn;
 
   /// Default duration animation
-  static const Duration animationDuration = Duration(milliseconds: 400);
+  static Duration animationDuration = const Duration(milliseconds: 400);
 
   /// Default duration animation fast
-  static const Duration animationDurationFast = Duration(milliseconds: 200);
+  static Duration animationDurationFast = const Duration(milliseconds: 200);
 
   /// Default duration animation low
-  static const Duration animationDurationLow = Duration(milliseconds: 800);
+  static Duration animationDurationLow = const Duration(milliseconds: 800);
 
   /// Default duration text field input end
-  static const Duration delayInputEnd = Duration(milliseconds: 1000);
+  static Duration delayInputEnd = const Duration(milliseconds: 1000);
 }