+++ bloc_architecture_widget/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Change logic input to mixin Handle
+- AppParams support change value
+
 ## 0.0.1
 
 - Initial Open Source release.